anavara/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-environ
django_filters
drf_yasg
django[argon2]
six
"django-phonenumber-field[phonenumbers]"
django-extensions